Add a task to import the GPG key for the target repository.
Add a task to set up the rpm package mirror.

Add a license and meta information for Ansible Galaxy.

Fix syntax error in meta/main.yml.

Fix the typo in the schema registry rpm repo key variable name.

Add a task to install the schema registry rpm.

Add a task for installing the jdk instead of using the role.